+++ testing/test-get.js
@@ -26,59 +26,65 @@ const evaluate = (title, result, expected) => {
   }
 };
 
-// Testing if error. get() without parameters
+// Testing get() without parameters. All keys at 'public' space (2 keys)
 result = efemem.get();
-evaluate(`get()`, result, false);
+evaluate(`get()`, result, true);
+console.log(`result: ${JSON.stringify(result, null, 2)}`);
 
-// Testing if error. get() with incorrect key (null)
+// Testing get() with key null. All keys at 'public' space (2 keys)
 result = efemem.get(null);
-evaluate(`get(null)`, result, false);
+evaluate(`get(null)`, result, true);
 
-// Testing if error. get() with incorrect key (no string name)
+// Testing get(). key 123 at 'public' space not found
 result = efemem.get(123);
-evaluate(`get(123)`, result, false);
+evaluate(`get(123)`, result, true);
 
-// Testing if error. get() with incorrect key (first character not alphabetical)
+// Testing get() key '1abc' at 'public' space not found
 result = efemem.get("1abc");
-evaluate(`get("1abc")`, result, false);
+evaluate(`get("1abc")`, result, true);
 
-// Testing if error. get() with incorrect key (illegal character)
+// Testing get() key 'maxValue!' at 'public' space not found
 result = efemem.get("maxValue!");
-evaluate(`get("maxValue!")`, result, false);
+evaluate(`get("maxValue!")`, result, true);
 
-// Testing if error. get() with incorrect key (More than 100 characters)
+// Testing get() a key not found
 result = efemem.get(
   "maxValueqweiutyqw573456y23478tyqweiutyqewi57832465897346tr87wqyetiuqweyt2347856234875yt6iequwyteiqwuty34896723"
 );
 evaluate(
   `get("maxValueqweiutyqw573456y23478tyqweiutyqewi57832465897346tr87wqyetiuqweyt2347856234875yt6iequwyteiqwuty34896723")`,
   result,
-  false
+  true
 );
 
-// Testing if error. get with incorrect space (first character not alphabetical)
+// Testing get() a key not found
 result = efemem.get("maxValue", "1abc");
-evaluate(`get("maxValue", "1abc")`, result, false);
+evaluate(`get("maxValue", "1abc")`, result, true);
 
-// Testing if error. get with incorrect space (illegal character)
+// Testing get() a key not found
 result = efemem.get("maxValue", "config!");
-evaluate(`get("maxValue", "config!")`, result, false);
+evaluate(`get("maxValue", "config!")`, result, true);
 
-// Testing if error. get with incorrect space (more than 24 characters)
+// Testing get() a key not found
 result = efemem.get("maxValue", 100, "configiu09olki87ujhy65tgf");
-evaluate(`get("maxValue", 100, "configiu09olki87ujhy65tgf")`, result, false);
+evaluate(`get("maxValue", 100, "configiu09olki87ujhy65tgf")`, result, true);
 
 // Testing if error. get with an existing key, but not in the 'public' space
 result = efemem.get("maxValue");
-evaluate(`get("maxValue")`, result, false);
+evaluate(`get("maxValue")`, result, true);
 
 // Testing a non existing key
 result = efemem.get("employee:007", "employees");
-evaluate(`get("employee:007", "employees")`, result, false);
+evaluate(`get("employee:007", "employees")`, result, true);
 
 // Testing an existing key
 result = efemem.get("employee:001", "employees");
 evaluate(`get("employee:001", "employees")`, result, true);
+
+// Testing an existing key with full info
+result = efemem.get("employee:001", "employees", true);
+evaluate(`get("employee:001", "employees", true)`, result, true);
+console.log(`result: ${JSON.stringify(result, null, 2)}`);
 
 console.log("\n-------------------------------------");
 console.log(`OK: ${ok} vs KO: ${ko}`);