+++ testing/test-move.js
@@ -26,14 +26,6 @@ const evaluate = (title, result, expected) => {
   }
 };
 
-// Testing if error. move() without parameters
-result = efemem.move();
-evaluate(`move()`, result, false);
-
-// Testing if error. move() with incorrect key (null)
-result = efemem.move(null);
-evaluate(`move(null)`, result, false);
-
 // Testing if error. move() with incorrect key (no string name)
 result = efemem.move(123);
 evaluate(`move(123)`, result, false);
@@ -84,9 +76,9 @@ evaluate(
   false
 );
 
-// Testing if error. move a valid key from an invalid space name (public by default) to a new valid space
-result = efemem.move("salary", null, "employees");
-evaluate(`move("salary", null, "employees")`, result, false);
+// If null origin space, assumes 'all' as default
+//result = efemem.move("salary", null, "employees");
+//evaluate(`move("salary", null, "employees")`, result, false);
 
 // Testing if error. move a valid key from an invalid space name to a new valid space
 result = efemem.move("salary", "config", "employees");
@@ -95,8 +87,25 @@ evaluate(`move("salary", "config", "employees")`, result, false);
 // Testing move a valid key from a valid space name to a new valid space (public by default)
 result = efemem.move("salary", "employees");
 evaluate(`move("salary", "employees")`, result, true);
-
 console.log(`result: ${JSON.stringify(result, null, 2)}`);
+
+// Checks move without overwrite
+efemem.set("maxValue", 200);
+// Try to move maxValue key from config to public space (existing) without overwriting
+result = efemem.move("maxValue", "config");
+evaluate(`move("maxValue", "config")`, result, true);
+console.log(`result: ${JSON.stringify(result, null, 2)}`);
+
+// Try to move maxValue key from config to public space (existing) overwriting
+result = efemem.move("maxValue", "config", null, true);
+evaluate(`move("maxValue", "config", null)`, result, true);
+console.log(`result: ${JSON.stringify(result, null, 2)}`);
+
+// ONLY FOR EXPLICITY TESTING
+// Moves all key to 'public' space
+//result = efemem.move();
+//evaluate(`move()`, result, true);
+//console.log(`result: ${JSON.stringify(result, null, 2)}`);
 
 console.log("\n-------------------------------------");
 console.log(`OK: ${ok} vs KO: ${ko}`);